dashboard: update windows-amd64-race builder to Windows Server 2016 image
The purpose of the -race builder is to check for bugs on the Go side,
so use the newer available Windows version with the expectation that
there should be fewer unfixed bugs left in it.

Take this easy step to eliminate a variable for issue golang/go#39196.

Also delete host-windows-amd64-2008-big, since it's no longer used.

For golang/go#39196.

Change-Id: I9095a73ab21a151e5972219bccd5a16b249765b4
Reviewed-on: https://go-review.googlesource.com/c/build/+/235419
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>