dashboard: document and clarify defaultBuildsRepoPolicy
The defaultBuildsRepoPolicy function was very important in that any
changes to it would affect many builders, but it was not documented
well.

This refactor attempts to improve the situation. It turned out better
to also rename defaultBuildsRepoPolicy to buildRepoByDefault and limit
its scope to report whether a given repo should be built by default.
The branch and goBranch parameters were never used, and thus misleading.
If they're needed in the future, we can easily re-add them.

The name defaultBuildsRepoPolicy made it sound as if it was the entire
"builds repo" policy, bun reality most of the policy is implemented in
the BuildConfig.buildsRepoAtAll method and is non-configurable. Only a
small part of the policy was in defaultBuildsRepoPolicy.

For golang/go#39156.

Change-Id: Ie7c6a68bd624056362dc7d964bd8d640f665b35a
Reviewed-on: https://go-review.googlesource.com/c/build/+/240687
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>