cmd/releasebot: remove obsolete checkDocs check
The check was added for golang/go#34045, back when doc/contrib.html
lived in the main repo and needed to be manually updated before each
major Go release. This is no longer applicable after CL 229483 and
CL 229485, so remove it before the upcoming Go 1.15 release.

This is compatible with future major releases and does not affect
upcoming minor releases of previous major releases.

Fixes golang/go#40476.
Updates golang/go#34045.

Change-Id: Ifc54dce106633e3d6b9ccdb0db4dae2b7434ce93
Reviewed-on: https://go-review.googlesource.com/c/build/+/245477
Reviewed-by: Carlos Amedee <carlos@golang.org>