dashboard: skip known missing ios/arm64 builder
We are aware that a builder is missing, and have the tracking issue
golang.org/issue/41610 for resolving that. Add a skip to the tests
so that the build dashboard is green until the work of adding the
builder is complete.

For golang/go#41610.

Change-Id: I69dbeb4c28d5567f24b4570f44715d8b5443365f
Reviewed-on: https://go-review.googlesource.com/c/build/+/257117
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>