internal/cloud: do not use canceled context in rate limiter
When errgroup is used in the rate limiter, a canceled context is
passed to the next method in the interceptor. A context is canceled
once the Wait() method returns. This change uses the same context
throughout the method instead of creating a new one just for the
errgroup.

For golang/go#40950

Change-Id: I64462394298a6b849187c288e5f76e92630572f9
Reviewed-on: https://go-review.googlesource.com/c/build/+/268697
Trust: Carlos Amedee <carlos@golang.org>
Run-TryBot: Carlos Amedee <carlos@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>