cmd/gopherbot: add task to ping "early-in-cycle" issues
There are two tasks we regularly do when reopening the tree for the
next major Go version development: remove "wait-release" hashtags
from CLs, and ping issues with "early-in-cycle" label.

The former was automated in CL 129817. The latter is automated here.
Maintner is used to find issues, and a GitHub API client is used to
post comments.

The Go 1.17 cycle was used as an opportunity to test this approach.
It worked as expected; no unexpected problems were found.

Like the unwait-release task, this is run manually via -only-run flag.
In the future we may want to factor both of these tasks out of
gopherbot and into a place that can be used by release automation.
But this seems like the most sensible place to add it initially.

Add a brief sleep in the hot loop for tasks that are invoked
manually. This is to appease humans and can modified at will.
It's unlikely to remain needed when these tasks are automated further.

For golang/go#40279.

Change-Id: I70c4eb7e98551d610f1518d1599661db21e2f271
Reviewed-on: https://go-review.googlesource.com/c/build/+/300570
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>