gerrit.ProjectInfo struct should include web_links
Gerrit sends web_links back in its response by default. But since
the struct type in the gerrit package doesn't have this, we end
up just ignoring it.

This change explicitly adds it to the ProjectInfo model so that
it gets de-serialized along with the rest of the response contents.