cmd/gitmirror: add tests
Add integration tests. They're on the slow side, as much as a second per
test, but the coverage is pretty solid and I hope they'll be easy to
read and maintain.

The tests run a real mirror, but without the normal driver loop and with
the repositories redirected to point to the local filesystem. That costs
us coverage of things like the gerrit polling, but removes the timing
sensitivity. I think that's a reasonable tradeoff; we'd just be mocking
maintner and gerrit anyway.

Also fix a bug that the tests surfaced: we can't run clone in the
directory we want it to create.

Change-Id: I8fa319b9bd3de194e75a2fdc2ee5a5cef71ec1aa
Reviewed-on: https://go-review.googlesource.com/c/build/+/324762
Trust: Heschi Kreinick <heschi@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>