cmd/retrybuilds: log failures that are being wiped
Saying nothing when a task is completed successfully is a good strategy
for some commands to use. However, it doesn't work well for retrybuilds,
since the command wipes matching failures, but it's not always known in
advance what the matching failures will be.

For similar reasons, add a -dry-run flag, to make it possible to run
the command in dry-run mode and confirm it will do the expected thing
before doing it for real.

For golang/go#34744.

Change-Id: I1517f79c6615bed9b297069daa83b53c927dc9bd
Reviewed-on: https://go-review.googlesource.com/c/build/+/336213
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>