internal/task, cmd/releasebot: add mail-dl-cl task to releasebot CLI
This change adds a release task that exercises the ability to create
a Gerrit CL via the Gerrit API, which will play a part in release
automation.

The task will eventually be exposed via relui's web UI when it is ready,
and it has a signature we expect to be able to use. To be able to use it
sooner, it's exposed via releasebot's CLI.

For golang.org/issue/38075.

Change-Id: I05ff078d7fcc25b1c2ac7001b094f3a3fee6b3a1
Reviewed-on: https://go-review.googlesource.com/c/build/+/350631
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Trust: Dmitri Shuralyov <dmitshur@golang.org>