cmd/release: start using macOS 12 releaselets in Go 1.18 Beta 1
The work to add macOS 12 builders for amd64 and arm64 was completed
ahead of Go 1.18 Beta 1, so we have the opportunity to pick between
starting to use them or stay with existing macOS 11 builders longer.

As part of go.dev/issue/40561 we discussed our available options and
opted to select macOS 12 releaselets for Go 1.18 on both architectures,
starting with first beta release. This CL implements that decision.

Tested:
  I ran 'release -target=darwin-amd64 -version=go1.18beta123 -rev=HEAD'
  locally and it completed successfully; all.bash release tests passed.

For golang/go#40561.

Change-Id: Ia80c4862a5137b5264c7a855452be25bddd79840
Reviewed-on: https://go-review.googlesource.com/c/build/+/369958
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>