cmd/coordinator: add link to known issue on builders page
The tooltip for builders with a known issue is displayed consistently
as of the last commit, but accessing a link in a tooltip is not easy.
It's reasonable to expect that known issues would be linked on the
builders page. Add a column with links there to make it so.

For golang/go#42405.

Change-Id: I5e915132f2ca05efc69bc76da1f25b897569be5d
Reviewed-on: https://go-review.googlesource.com/c/build/+/370335
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>