dashboard: build x/vulndb only on linux-amd64
The vulndb repo is not intended for external use. It consists of some
binaries and servers used by the Go team for vulnerability work.  So
there is no need for it to build on a variety of systems and Go
versions.  In fact, it only needs to work on the current version of Go
on vanilla Linux.

Fixes golang/go#50472

Change-Id: I37bc0136ad92f2b959ff8ef4e89c3bde3a2d31d6
Reviewed-on: https://go-review.googlesource.com/c/build/+/375957
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>