internal/releasetargets: create
We duplicate a list of release targets between cmd/releasebot and
cmd/release. The format of the latter is, to me, unpleasant to deal
with. This is an attempt at a better one, more information-dense and
less error-prone.

AFAIK, the only thing that really matters is what builder we should use
to issue the next release in each series. Reproducibility of prior
releases is not so important, especially because when we change builders
mid-stream it's usually involuntary. So, rather than supporting the
full-fledged and quite complicated GoQuery syntax, simply have an entry
per live release series.

Make the minimal changes to integrate the new package into cmd/release
and cmd/releasebot.

For golang/go#51797.

Change-Id: Ic48b1cdd751fb4ea5e2acad8f1898b48faa30cd9
Reviewed-on: https://go-review.googlesource.com/c/build/+/387714
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Alex Rakoczy <alex@golang.org>