cmd/coordinator: clean up branch-related logic in buildStatus
In service of getting the right branch to performance monitoring data,
this change stops using the confusingly named (and confusingly set)
goBranch field of buildStatus in favor of always passing a commitDetail
to newBuild with a valid branch. This means that all uses of
commitDetail pass a non-nil commitDetail, so this change also makes it
passed by value.

This change also does some additional quality-of-life refactoring.
Namely, it lifts the parsing of commit times into the commitDetail,
allowing us to just embed a commitDetail into the buildStatus struct.
It also removes dead code related to staging that would have to change
as a result of this CL.

This change is almost a no-op, except for performance monitoring
use-cases. In all cases where goBranch was used, the equivalent
RevBranch is now populated and used instead. For performance monitoring
however, by using RevBranch instead of goBranch, we guarantee that we
always have a branch and it's the one we actually want.

Conveniently, the branch is now always available to buildStatus when it
wasn't before, which may be useful in the future.

For golang/go#48803.

Change-Id: I8ad3f115b1b2546dbf0848649965b2c720f73ead
Reviewed-on: https://go-review.googlesource.com/c/build/+/395174
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Trust: Michael Knyszek <mknyszek@google.com>