cmd/release: move extra CFLAGS into releasetargets
We only need the one value for now, but it seems like it's going to
stick around. Rather than special-casing it, move it into the release
targets where it can be tracked along with the rest of the target
configuration.

For golang/go#51797.

Change-Id: I01f1f11bac3e0ea9a91ac4efc541e596004ea7d5
Reviewed-on: https://go-review.googlesource.com/c/build/+/395536
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Alex Rakoczy <alex@golang.org>