internal/relui: log previous state when retrying tasks
This logs the task state and workflow state when retrying a task for a
workflow, so as not to lose history.

Eventually, we should store the full history for workflows and tasks.

Moves the transaction rollback to a defer, which is safe even if commit
or rollback is called first.

Fixes golang/go#53165

Change-Id: I6a9c42beb9352656b044dbe6e5adf6493d32873d
Reviewed-on: https://go-review.googlesource.com/c/build/+/411194
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Alex Rakoczy <alex@golang.org>