internal/relui: add approval step to build workflow
This change adds a manual approval task to the build before making any
user-facing changes. It uses a sentinel log entry in the approval task's
log in order to verify whether a human has approved the workflow to
continue.

Ideally, we'd change the data model to allow for this, perhaps by adding
a specific type of task, or a special state column, but this should be
sufficient for testing the beta release. Also, there should be some
shared definition of approval, but leaving that for a follow-up CL.

Updates golang/go#53295
Updates golang/go#51797

Change-Id: Iacd8c8a3ba59f9e3a343916dbe1a3a8ba59ce80c
Reviewed-on: https://go-review.googlesource.com/c/build/+/411199
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Alex Rakoczy <alex@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>