internal/task: add release announcement email templates and test data
This CL contains just the templates and test data for review
convenience. The next CL in the stack holds the implementation.

The email templates have a MIME-style header with a Subject key,
which is used to determine the email subject, and the rest is
the email body in Markdown format.

While the announcement emails have been generated from templates
for a while, there was variation in minor formatting and wording
over the years. These templates are converted to Markdown and
attempt to not to deviate from the average past announcements
too much while correcting minor inconsistencies.

Making larger changes can be considered later. It should be easier
now that this task is driven by templates without outsourcing the
responsibility of HTML formatting and editing to the human senders.

For golang/go#47405.
Fixes golang/go#47404.

Change-Id: Ie04d94b35ce1b5d111eeb1b0b076c034ba3f4dcb
Reviewed-on: https://go-review.googlesource.com/c/build/+/412754
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>