cmd/gopherbot: add task to label compiler/runtime issues
This change adds a task to gopherbot to automatically label
compiler/runtime issues based on the owners list. This task will aid in
finding compiler/runtime issues for further triaging, prioritization,
and assigning ownership.

This change also adds a JSON field to the code owners API to allow for
obtaining all ownership information. The reason why an explicit field
was chosen over alternatives like an empty path set meaning 'all' is
because it's backwards compatible.

Finally, it exports the translation that the owners page does for
package names so that gopherbot can translate the paths the owners
page returns into the names used for packages in GitHub issues.

Change-Id: Idb81a2ab53d16f8bec71450899ea33a5e1c30abe
Reviewed-on: https://go-review.googlesource.com/c/build/+/413579
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>